Перенос Cell в maps
Импорт Cell из game и одновременный импорт nodes_matrix из maps не срабатывает, поэтому было решено перенести Cell в maps.
