+++ maps.py
@@ -1,5 +1,33 @@
-from game import Cell
-from typing import List
+from typing import List, Union
+
+
+class Cell:
+    def __init__(self, x, y, cell_type: Union[int, str] = 0, food: int = 1):
+        '''cell_type: 0 или field для поля, 1 или wall для стены, по умолчанию 0\n
+        food: для еды 1, для энергии 10, по умолчанию 1, если клетка является полем'''
+        self.x, self.y = x, y  # Координаты клетки
+        self.previous = None  # Предыдущий оптимальный по пути узел графа.
+        self.cost = 58  # Цена узла графа. Необходима для расчёта
+        if isinstance(cell_type, str):
+            self.type = cell_type  # field для ячейки, по которой можно ходить, и wall для ячейки, являющейся стеной.
+        else:
+            self.type = 'field' if cell_type == 0 else 'wall'
+
+        # Думаю, как реализовать наличие еды в клетке. Для self.type == 'wall' можно определять None.
+        self.has_food = (True if food == 1 else False) if self.type != 'wall' else None
+        # Думаю, как реализовать наличие энерджайзера в клетке. Для self.type == 'wall' можно определять None.
+        self.has_energy = (True if food == 10 else False) if self.type != 'wall' else None
+
+    def reset(self):
+        # Сброс характеристик клетки для построения следующего пути.
+        self.previous = None
+        self.cost = 58
+
+    def __repr__(self) -> str:
+        return ('0' if self.x < 10 else '') + str(self.x) + '-' + ('0' if self.y < 10 else '') + str(self.y)
+
+    def __str__(self) -> str:
+        return '-' if self.type == 'wall' else ('E' if self.has_energy else ('F' if self.has_food else '0'))
 
 
 r = range
